refactor(dashboards): extract useTopologyData hook from NetworkDashboard
Phase 1 of #1071's scoping plan. NetworkDashboard.tsx's data layer
(graph fetch + twin-driven auto-refresh + the two effects that drive
them) moves into packages/frontend/src/hooks/useTopologyData.ts. UI
concerns (toast plumbing, selection state, layout, modals) all stay in
the component; the hook is a pure data layer with two pass-through
callbacks (onLoadStart, onLoadError) so it stays testable and has no
toast-provider dependency.

NetworkDashboard.tsx drops ~50 LOC of fetch wiring + a stale "compute
graph from Twin on the fly" comment block (the compromise it documented
landed in the hook's auto-fetch effect). The component still owns
rawGraphData, the layout pipeline, and the selection state — those are
Phase 2 (NodeDetailsDrawer) and Phase 3 (TopologyViewer) per the
scoping plan; this PR is intentionally narrow.

File line count: 2083 → 2062. Lint warning count unchanged at 428;
the file's max-lines warning ratchets down (1762 → 1741) and the
NetworkDashboard function's line count drops accordingly.

Phase 1 of #1071. Issue stays open for Phase 2/3 follow-ups.
